requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
decorator==5.1.1
Django==4.2.2
django-cors-headers==4.1.0
django-password-validators==1.7.1
django-rest-framework==0.1.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
dnspython==2.3.0
dynaconf==3.1.12
email-validator==2.0.0.post2
hvac==1.1.1
idna==3.4
loguru==0.7.0
pycparser==2.21
pyhcl==0.4.4
pytz==2023.3
requests==2.31.0
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.3
validators==0.20.0
